backend/function-food-chat/requirements.txt
functions-framework==3.5.0
google-genai==0.1.0
google-cloud-storage==2.14.0
Flask==3.0.0
